requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
PyMySQL==1.1.1
Flask-WTF==1.2.1
email-validator==2.2.0
python-dotenv==1.0.1
blinker==1.9.0
cffi==2.0.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dnspython==2.8.0
email_validator==2.2.0
Flask==3.0.3
Flask-WTF==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.23
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
